jabber: fix many conversion issues

+++ plugins/jabber_protocol/gui/widgets/jabber-create-account-widget.h
@@ -75,7 +75,7 @@ class JabberCreateAccountWidget : public AccountCreateWidget
 	int ssl_;
 	bool opt_host_, legacy_ssl_probe_;
 	QString host_;
-	quint16 port_;
+	uint port_;
 	bool ShowConnectionOptions;
 
 	void createGui(bool showButtons);
